refactor: Extract useUser hook
