+++ app/login/components/login-menu.tsx
@@ -1,48 +1,18 @@
 import { Box, Button, Typography } from "@mui/material";
-import { getProviders, signIn, useSession } from "next-auth/react";
+import { signIn } from "next-auth/react";
 import Link from "next/link";
-import { useEffect, useState } from "react";
 
-import { Awaited } from "@/domain/types";
-
-type Providers = Awaited<ReturnType<typeof getProviders>>;
-
-const useProviders = () => {
-  const [state, setState] = useState({
-    status: "loading",
-    data: undefined as Providers | undefined,
-  });
-
-  useEffect(() => {
-    const run = async () => {
-      const providers = await getProviders();
-      setState({ status: "loaded", data: providers });
-    };
-
-    run();
-  }, []);
-
-  return state;
-};
+import { useUser } from "@/login/utils";
 
 export const LoginMenu = () => {
-  const { data: session, status: sessionStatus } = useSession();
-  const { data: providers, status: providersStatus } = useProviders();
-
-  if (sessionStatus === "loading" || providersStatus === "loading") {
-    return null;
-  }
-
-  if (!providers || !Object.keys(providers).length) {
-    return null;
-  }
+  const user = useUser();
 
   return (
     <Box sx={{ alignItems: "center", display: "flex" }}>
-      {session ? (
+      {user ? (
         <Typography variant="body2">
           <Link href="/profile" legacyBehavior>
-            {session.user?.name}
+            {user.name}
           </Link>{" "}
         </Typography>
       ) : (
